fix(dashboard): reset stranded in-flight flags on disconnect (#6285, #6289)
De-scoped to the two clean flag-reset fixes; #6287 (permission-keyboard
single-listener) split into its own PR — it moves keyboard handling out of
PermissionPrompt and needs a careful test migration that warrants focused,
security-sensitive review.

#6285 — createSession returns a boolean (true=sent); App only latches the
'Creating…' spinner when the request went on the wire, and a new effect clears
it + surfaces a retryable error if connectionPhase leaves 'connected' mid-create
(was: spinner wedged forever on a socket drop between Create and reply).

#6289 — externalSessionsLoading added to the #6153 surveyLoadingKeys onclose
sweep (the snapshot handler already cleared it on success; only the drop-path
reset was missing) — closes the latent refresh-deadlock once a Refresh is wired.
